Translate click event on client to grid coordinate

+++ socket/request.go
@@ -2,6 +2,6 @@ package socket
 
 type Request struct {
 	Type string `json:type`
-	X	float64	`json:x`
-	Y	float64	`json:y`
+	X	int	`json:x`
+	Y	int	`json:y`
 }
